% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{team_conference_logos}
\alias{team_conference_logos}
\title{Team and Conference Logos}
\format{
A data frame with 155 rows and 6 variables:
\describe{
\item{team}{the long-form name of the school (e.g., "University of California, Berkeley" for "Cal")}
\item{team_short}{a lowercase, one-word, shortened school identifier (e.g., "cal" for "Cal")}
\item{conference}{the abbreviated name of the conference (e.g., "SEC" for "Southeastern Conference")}
\item{conference_short}{a lowercase, one-word, abbreviated name for the conference}
\item{team_link}{the url for the team logo}
\item{conference_link}{the url for the conference logo}
}
}
\usage{
team_conference_logos
}
\description{
A dataset containing links to team and conference logos. The variables are:
}
\keyword{datasets}
