% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{vepv_add_opponent}
\alias{vepv_add_opponent}
\title{Adds a column indicating the opposing team to a dataset of plays}
\usage{
vepv_add_opponent(plays)
}
\arguments{
\item{plays}{the data frame containing the plays}
}
\value{
the dataset with a new column \code{opponent}
}
\description{
Adds a column indicating the opposing team to a dataset of plays
}
