% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vepv_add_positions}
\alias{vepv_add_positions}
\title{Add positions}
\usage{
vepv_add_positions(plays, position_df = NULL)
}
\arguments{
\item{plays}{the data frame containing the plays}

\item{position_df}{a data frame with two columns: \code{player_name}, the name of the player (as given in \code{plays}), and the player's position.
If \code{NULL}, or for any players not found in \code{position_df}, a position will be guessed based on the types of touches in the \code{plays} data}
}
\value{
the same data frame, with a \code{position} column indicating the actual (or guessed) position of the player
}
\description{
Adds a position for each player. Will guess the position if none is provided.
}
