% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{vepv_add_possession_contacts}
\alias{vepv_add_possession_contacts}
\title{Possession and Contact Numbers}
\usage{
vepv_add_possession_contacts(plays)
}
\arguments{
\item{plays}{the data frame containing the plays}
}
\value{
the dataset with new \code{skq} column for skill-quality combination
}
\description{
Adds possession numbers and contact number within possession
}
\details{
Add a skill-quality combination column to a dataset and recode problematic skills, e.g., add a "Cover" skill
}
