% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{vepv_add_rally_winner}
\alias{vepv_add_rally_winner}
\title{Add rally winner}
\usage{
vepv_add_rally_winner(plays)
}
\arguments{
\item{plays}{the data frame containing the plays}
}
\value{
the same data frame with \code{rally_winner} and \code{rally_eff} columns. \code{rally_winner} indicates whether the touching team won the point (1) or not (0),
and \code{rally_eff} converts this to an efficiency (1 or -1)
}
\description{
Adds a variable indicating whether the point was won by the team touching the ball
}
