% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{vepv_add_touch_id}
\alias{vepv_add_touch_id}
\title{Add Touch ID}
\usage{
vepv_add_touch_id(plays)
}
\arguments{
\item{plays}{the data frame containing the plays}
}
\value{
the dataset with a new column \code{id_touch}
}
\description{
Adds a column indicating the row number in the original dataset of plays
}
