% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{vepv_add_year}
\alias{vepv_add_year}
\title{Add year}
\usage{
vepv_add_year(plays)
}
\arguments{
\item{plays}{the data frame containing the plays}
}
\value{
the dataset with a new column \code{year}
}
\description{
Adds a column indicating the match year to a dataset of plays
}
