% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{vepv_blockers}
\alias{vepv_blockers}
\title{Add number of blockers}
\usage{
vepv_blockers(
  plays,
  impute = TRUE,
  impute_seed = 100,
  impute_probs = data.frame(types = c("double", "solo", "triple", "seam", "none"), yes =
    c(0.48, 0.23, 0.224, 0.016, 0.007), no = c(0.724, 0.097, 0.063, 0.075, 0.012))
)
}
\arguments{
\item{plays}{the data frame containing the plays}

\item{impute}{a logical variable indicating whether to impute (TRUE) or not (FALSE) a value for the number of blockers when it is missing}

\item{impute_seed}{the seed to use for the imputation}

\item{impute_probs}{a data frame containing three columns:
types of blocks, probability of facing each type of block when the attack has a k-code and probability of facing each type of block when the attack does not have a k-code}
}
\value{
the same data frame with \code{input_type} and \code{output_type} columns
}
\description{
Adds the number of blockers
}
