% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{vepv_data_processing_workflow}
\alias{vepv_data_processing_workflow}
\title{Expected Point Value Data Processing Workflow}
\usage{
vepv_data_processing_workflow(folder)
}
\arguments{
\item{folder}{the file path to the folder containing the .dvw files to process}
}
\value{
a data frame containing the processed data from the .dvw files
}
\description{
Runs all data processing functions in the correct order to get the data in a form for modeling.
}
