% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{vepv_dig_difficulty}
\alias{vepv_dig_difficulty}
\title{Add dig difficulty}
\usage{
vepv_dig_difficulty(plays)
}
\arguments{
\item{plays}{the data frame containing the plays}
}
\value{
the same data frame with \code{speed}, \code{block_touch}, and \code{reaction_time} columns. \code{speed} represents whether the previous attack was a hard spike or not,
\code{block_touch} indicates whether the dig was directly off an attack or off a block-touch,
and \code{reaction_time} indicates the (approximate) time between the previous touch and the dig
}
\description{
Adds variables related to the difficulty of the dig
}
