% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vepv_epv}
\alias{vepv_epv}
\title{Expected Point Value}
\usage{
vepv_epv(plays, touch_epv = NULL)
}
\arguments{
\item{plays}{the data frame containing the plays}

\item{touch_epv}{a data frame with two columns: \code{touch}, the type of touch (as indicated by\code{input_type} or \code{output_type}), and the expected value of that touch to impute.
If \code{NULL}, or for any touch types not found in \code{touch_epv}, a value will be imputed using the \code{plays} data}
}
\value{
a data frame containing input, output, and added expected point value for each touch
}
\description{
Calculates expected rally efficiency for each touch
}
