% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{vepv_full_workflow}
\alias{vepv_full_workflow}
\title{Expected Point Value Full Workflow}
\usage{
vepv_full_workflow(folder, touch_file = NULL, position_file = NULL)
}
\arguments{
\item{folder}{the file path to the folder containing the .dvw files to process}

\item{touch_file}{optionally, either the name of an Excel or CSV file containing the data to pass on to the modeling functions as \code{touch_epv}, or the data frame itself.
The default (= \code{NULL}) is to not give a file and have all values imputed using the data from the .dvw files.}

\item{position_file}{either the name of an Excel or CSV file containing the data to pass on to the modeling functions as \code{position_df}, or the data frame itself
The default (= \code{NULL}) is to not give a file and have all values imputed using the data from the .dvw files.}
}
\value{
a data frame containing the processed data from the .dvw files
}
\description{
Runs all data processing and modeling functions in the correct order to get the data in a form for modeling.
}
