% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{vepv_import_files}
\alias{vepv_import_files}
\title{Import DataVolley Files}
\usage{
vepv_import_files(file_list, verbose = TRUE)
}
\arguments{
\item{file_list}{a vector of file names with .dvw extensions}

\item{verbose}{a logical indicating whether to print out the cumulative number of files imported}
}
\value{
a data frame containing the plays objects of each file in the set
}
\description{
Extracts the plays objects from a set of .dvw files and concatenates into one data frame
}
