% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vepv_input_location_model}
\alias{vepv_input_location_model}
\title{Input and Output Models}
\usage{
vepv_input_location_model(plays, touch_type, trained_model = NULL)
}
\arguments{
\item{plays}{the data frame containing the plays}

\item{touch_type}{the type of touch (as indicated by the \code{input_type} or \code{output_type} column of \code{plays})}

\item{trained_model}{a pre-trained model (of class "\code{glm}") used to make predictions}
}
\value{
the proportion of these touches that eventually resulted in winning the point
}
\description{
Calculates expected rally efficiency using a logistic regression model, which depends on the touch type
}
