% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vepv_mean_model}
\alias{vepv_mean_model}
\title{Simple mean model}
\usage{
vepv_mean_model(plays, touch_type, impute_value = NA_real_)
}
\arguments{
\item{plays}{the data frame containing the plays}

\item{touch_type}{the type of touch (as indicated by the \code{input_type} column of \code{plays})}

\item{impute_value}{the value to impute for the rally efficiency for this type of touch, if not estimated directly from data}
}
\value{
the proportion of these touches that eventually resulted in winning the point
}
\description{
Calculates expected rally efficiency using a simple proportion of these types of touches that result in winning the point.
Useful for touches that either don't have useful predictors or can't make use of them because they're coded weird.
}
