% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{vepv_modeling_workflow}
\alias{vepv_modeling_workflow}
\title{Expected Point Value Modeling Workflow}
\usage{
vepv_modeling_workflow(plays, touch_epv = NULL, position_df = NULL)
}
\arguments{
\item{plays}{the data frame containing the plays}

\item{touch_epv}{a data frame with two columns: \code{touch}, the type of touch (as indicated by\code{input_type} or \code{output_type}), and the expected value of that touch to impute.
If \code{NULL}, or for any touch types not found in \code{touch_epv}, a value will be imputed using the \code{plays} data}

\item{position_df}{a data frame with two columns: \code{player_name}, the name of the player (as given in \code{plays}), and the player's position.
If \code{NULL}, or for any players not found in \code{position_df}, a position will be guessed based on the types of touches in the \code{plays} data}
}
\value{
a data frame containing input, output, and added expected point value for each touch, and the (possibly guessed) position for each player
}
\description{
Runs all modeling functions in the correct order to get the data in a form for modeling.
}
