% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{vepv_recode_skills}
\alias{vepv_recode_skills}
\title{Recode skills}
\usage{
vepv_recode_skills(plays)
}
\arguments{
\item{plays}{the data frame containing the plays}
}
\value{
the dataset with updated \code{evaluation_code} and \code{skill} columns and a new \code{skq} column for skill-quality combination
}
\description{
Add a skill-quality combination column to a dataset and recode problematic skills, e.g., add a "Cover" skill
}
