% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{vepv_remove_nonactions}
\alias{vepv_remove_nonactions}
\title{Remove Non-Actions}
\usage{
vepv_remove_nonactions(plays)
}
\arguments{
\item{plays}{the data frame containing the plays}
}
\value{
a subset of the data frame containing only the touches
}
\description{
Removes the non-actions from a data frame containing plays objects
}
