% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{vepv_touch_coordinates}
\alias{vepv_touch_coordinates}
\title{Add touch coordinates}
\usage{
vepv_touch_coordinates(plays)
}
\arguments{
\item{plays}{the data frame containing the plays}
}
\value{
the same data frame with \code{input_coord_x}, \code{input_coord_y}, \code{output_coord_x}, and \code{output_coord_y} columns
}
\description{
Adds the input and output x and y-coordinates for the touch
}
