% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{vepv_touch_input_output}
\alias{vepv_touch_input_output}
\title{Add touch inputs and outputs}
\usage{
vepv_touch_input_output(plays)
}
\arguments{
\item{plays}{the data frame containing the plays}
}
\value{
the same data frame with \code{input_type} and \code{output_type} columns
}
\description{
Adds the touch input and the touch result (output)
}
