% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{vepv_touch_kcode}
\alias{vepv_touch_kcode}
\title{Add k-code}
\usage{
vepv_touch_kcode(plays)
}
\arguments{
\item{plays}{the data frame containing the plays}
}
\value{
the same data frame with \code{input_kcode} and \code{output_kcode} columns
}
\description{
Adds a variable indicating whether the touch leads to/comes from a set with a k-code
}
